refactor: use double instead of float
Jackson Ion does not support serializing floats correctly, so we'll just use doubles

+++ lib/src/main/java/sharetrace/algorithm/Contact.java
@@ -19,7 +19,7 @@ final class Contact implements Expirable, Comparable<Contact> {
   private final Instant timestamp;
   private final Instant bufferedTimestamp;
   private final Instant expiresAt;
-  private final float sendCoefficient;
+  private final double sendCoefficient;
   private final Cache<?, RiskScoreMessage> scores;
   private final InstantSource timeSource;
